Atom color array filling is now a rendering pre-pass executed as required
